Support spring
This introduces the `check_for_bundler` flag to be set in the settings file. If set to true, it will prepend the `COMMAND` with the [`spring`](https://github.com/jonleighton/spring) command.

NOTE that the documentation for spring recommends that you don't use spring from inside Bundler. So you should turn off bundler support if turning on spring support.